feat: migrate to next.js pages router and fixed re-rendering issue
