+++ next.config.js
@@ -3,7 +3,6 @@ const nextConfig = {
     experimental: {
         serverActions: true
     },
-    compress: true,
     compiler: {
         removeConsole: process.env.NODE_ENV !== 'development'
     },
