+++ package.json
@@ -16,14 +16,14 @@
     "autoprefixer": "10.4.14",
     "clsx": "^1.2.1",
     "eslint": "8.40.0",
-    "eslint-config-next": "13.4.2",
+    "eslint-config-next": "^13.4.5",
     "framer-motion": "^10.12.16",
     "html2canvas": "^1.4.1",
     "lucide-react": "^0.236.0",
-    "next": "^13.4.4-canary.13",
+    "next": "^13.4.5",
     "postcss": "8.4.23",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-wrap-balancer": "^0.5.0",
     "sharp": "^0.32.1",
     "swr": "^2.1.5",
