--- src/app/image/layout.tsx
@@ -1,9 +1,0 @@
-export default function ImageLayout({ children }: {
-    children: React.ReactNode
-}) {
-    return (
-        <div className="p-5 w-screen h-screen flex flex-col items-center justify-center">
-            {children}
-        </div>
-    )
-}
