--- src/app/image/loading.tsx
@@ -1,3 +1,0 @@
-import Loader from "@/components/Loader";
-
-export default Loader;
